// src/types/RockGenres.ts
export enum RockGenres {
    goticrock = 0,
    hardrock = 1,
    punkrock = 2,
    hardcore = 3,
    deathcore = 4,
    metalcore = 5,
    heavymetal = 6,
    glammetal = 7,
    goticmetal = 8,
    blackmetal = 9,
    paganmetal = 10,
    symphonicblackmetal = 11,
    deathmetal = 12,
    symphonicdeathmetal = 13,
}